nixos/pam: automatically populate rule type
Eliminates a redundancy between the type attribute key and the type
specified in each rule.

Why introduce the type attribute in the first place? Because we
eventually want to give each rule a name so that we can merge config
overrides. The PAM name is a natural choice for rule name, but a PAM is
often used in multiple rule types. Organizing rules by type and rule
name avoids name collisions.