nixos/pam: convert rules to attrs with order field
In order for a user to override properties of a rule, we make rules
into an attrset with predictable keys. We also introduce an ordering
field.

For now, the order value for each built-in rule is derived from its
place in the hardcoded list of rules. Since a change to the rules list
would change the order values of multiple rules, this option is
potentially hazardous. It is marked experimental and not visible.